requirements.txt
aiogram>=3.7,<4
aiohttp>=3.9,<4
yookassa>=3.2,<4
python-dotenv>=1.0,<2
asyncpg>=0.29,<1
